Allow Powered by Death mutation to work for Gozag.
Gozagites don't get corpses, so for the purpose of this mutation use
gold stacks instead for the stacking bonus. This is somewhat problematic
in that unlike corpses, gold stacks and never expires, which could
somewhat encourage degenerate behaviour. Hopefully it's balanced by PbD
being significantly less effective when not fighting out in the open
(you won't get PbD stacks in a corridor).